+++ proto/src/prost/tendermint.crypto.rs
@@ -54,7 +54,7 @@ pub struct ProofOps {
 #[derive(Clone, PartialEq, ::prost::Message)]
 #[derive(::serde::Deserialize, ::serde::Serialize)]
 pub struct PublicKey {
-    #[prost(oneof="public_key::Sum", tags="1")]
+    #[prost(oneof="public_key::Sum", tags="1, 2")]
     pub sum: ::std::option::Option<public_key::Sum>,
 }
 pub mod public_key {
@@ -65,5 +65,8 @@ pub mod public_key {
         #[prost(bytes, tag="1")]
         #[serde(rename = "tendermint/PubKeyEd25519", with = "crate::serializers::bytes::base64string")]
         Ed25519(std::vec::Vec<u8>),
+        #[prost(bytes, tag="2")]
+        #[serde(rename = "tendermint/PubKeySecp256k1", with = "crate::serializers::bytes::base64string")]
+        Secp256k1(std::vec::Vec<u8>),
     }
 }