+++ tendermint/src/abci/transaction.rs
@@ -95,10 +95,7 @@ impl From<RawData> for Data {
 impl From<Data> for RawData {
     fn from(value: Data) -> Self {
         if value.txs.is_none() {
-            return RawData {
-                txs: vec![],
-                hash: vec![],
-            };
+            return RawData { txs: vec![] };
         }
         RawData {
             txs: value
@@ -107,7 +104,6 @@ impl From<Data> for RawData {
                 .iter()
                 .map(|tx| tx.clone().into())
                 .collect(),
-            hash: vec![],
         }
     }
 }