+++ tendermint/src/error.rs
@@ -115,6 +115,10 @@ pub enum Kind {
     #[error("missing evidence field")]
     MissingEvidence,
 
+    /// Missing Timestamp in Block
+    #[error("missing timestamp field")]
+    MissingTimestamp,
+
     /// Invalid Block
     #[error("invalid block")]
     InvalidBlock,