+++ tendermint/src/evidence.rs
@@ -1,6 +1,8 @@
 //! Evidence of malfeasance by validators (i.e. signing conflicting votes).
 
-use crate::{block::signed_header::SignedHeader, serializers, Error, Kind, Vote};
+use crate::{
+    block::signed_header::SignedHeader, serializers, vote::Power, Error, Kind, Time, Vote,
+};
 use serde::{Deserialize, Serialize};
 use std::convert::{TryFrom, TryInto};
 use std::slice;
@@ -9,7 +11,7 @@ use tendermint_proto::types::evidence::Sum as RawSum;
 use tendermint_proto::types::evidence::Sum;
 use tendermint_proto::types::DuplicateVoteEvidence as RawDuplicateVoteEvidence;
 use tendermint_proto::types::Evidence as RawEvidence;
-use tendermint_proto::types::EvidenceData as RawEvidenceData;
+use tendermint_proto::types::EvidenceList as RawEvidenceList;
 use tendermint_proto::types::EvidenceParams as RawEvidenceParams;
 use tendermint_proto::Protobuf;
 
@@ -62,6 +64,9 @@ impl From<Evidence> for RawEvidence {
 pub struct DuplicateVoteEvidence {
     vote_a: Vote,
     vote_b: Vote,
+    total_voting_power: Power,
+    validator_power: Power,
+    timestamp: Time,
 }
 
 impl TryFrom<RawDuplicateVoteEvidence> for DuplicateVoteEvidence {
@@ -71,6 +76,9 @@ impl TryFrom<RawDuplicateVoteEvidence> for DuplicateVoteEvidence {
         Ok(Self {
             vote_a: value.vote_a.ok_or(Kind::MissingEvidence)?.try_into()?,
             vote_b: value.vote_b.ok_or(Kind::MissingEvidence)?.try_into()?,
+            total_voting_power: value.total_voting_power.try_into()?,
+            validator_power: value.validator_power.try_into()?,
+            timestamp: value.timestamp.ok_or(Kind::MissingTimestamp)?.try_into()?,
         })
     }
 }
@@ -80,6 +88,9 @@ impl From<DuplicateVoteEvidence> for RawDuplicateVoteEvidence {
         RawDuplicateVoteEvidence {
             vote_a: Some(value.vote_a.into()),
             vote_b: Some(value.vote_b.into()),
+            total_voting_power: value.total_voting_power.into(),
+            validator_power: value.total_voting_power.into(),
+            timestamp: Some(value.timestamp.into()),
         }
     }
 }
@@ -91,7 +102,13 @@ impl DuplicateVoteEvidence {
             return Err(Kind::InvalidEvidence.into());
         }
         // Todo: make more assumptions about what is considered a valid evidence for duplicate vote
-        Ok(Self { vote_a, vote_b })
+        Ok(Self {
+            vote_a,
+            vote_b,
+            total_voting_power: Default::default(),
+            validator_power: Default::default(),
+            timestamp: Time::now(),
+        })
     }
     /// Get votes
     pub fn votes(&self) -> (&Vote, &Vote) {
@@ -120,14 +137,14 @@ impl ConflictingHeadersEvidence {
 ///
 /// <https://github.com/tendermint/spec/blob/d46cd7f573a2c6a2399fcab2cde981330aa63f37/spec/core/data_structures.md#evidencedata>
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-#[serde(try_from = "RawEvidenceData", into = "RawEvidenceData")]
+#[serde(try_from = "RawEvidenceList", into = "RawEvidenceList")]
 pub struct Data {
     evidence: Option<Vec<Evidence>>,
 }
 
-impl TryFrom<RawEvidenceData> for Data {
+impl TryFrom<RawEvidenceList> for Data {
     type Error = Error;
-    fn try_from(value: RawEvidenceData) -> Result<Self, Self::Error> {
+    fn try_from(value: RawEvidenceList) -> Result<Self, Self::Error> {
         if value.evidence.is_empty() {
             return Ok(Self { evidence: None });
         }
@@ -139,16 +156,15 @@ impl TryFrom<RawEvidenceData> for Data {
     }
 }
 
-impl From<Data> for RawEvidenceData {
+impl From<Data> for RawEvidenceList {
     fn from(value: Data) -> Self {
-        RawEvidenceData {
+        RawEvidenceList {
             evidence: value
                 .evidence
                 .unwrap_or_default()
                 .into_iter()
                 .map(Into::into)
                 .collect(),
-            hash: vec![],
         }
     }
 }
@@ -183,6 +199,8 @@ impl AsRef<[Evidence]> for Data {
 
 /// Evidence collection parameters
 #[derive(Deserialize, Serialize, Clone, Debug, Eq, PartialEq)]
+// Todo: This struct is ready to be converted through tendermint_proto::types::EvidenceParams.
+// https://github.com/informalsystems/tendermint-rs/issues/741
 pub struct Params {
     /// Maximum allowed age for evidence to be collected
     #[serde(with = "serializers::from_str")]
@@ -192,8 +210,8 @@ pub struct Params {
     pub max_age_duration: Duration,
 
     /// Max bytes
-    #[serde(default)]
-    pub max_num: i64,
+    #[serde(with = "serializers::from_str", default)]
+    pub max_bytes: i64,
 }
 
 impl Protobuf<RawEvidenceParams> for Params {}
@@ -211,7 +229,7 @@ impl TryFrom<RawEvidenceParams> for Params {
                 .max_age_duration
                 .ok_or(Kind::MissingMaxAgeDuration)?
                 .try_into()?,
-            max_num: value.max_num as i64,
+            max_bytes: value.max_bytes,
         })
     }
 }
@@ -222,7 +240,7 @@ impl From<Params> for RawEvidenceParams {
             // Todo: Implement proper domain types so this becomes infallible
             max_age_num_blocks: value.max_age_num_blocks.try_into().unwrap(),
             max_age_duration: Some(value.max_age_duration.into()),
-            max_num: value.max_num as u32,
+            max_bytes: value.max_bytes,
         }
     }
 }
@@ -231,6 +249,7 @@ impl From<Params> for RawEvidenceParams {
 /// essentially, to keep the usages look cleaner
 /// i.e. you can avoid using serde annotations everywhere
 /// Todo: harmonize google::protobuf::Duration, std::time::Duration and this. Too many structs.
+/// https://github.com/informalsystems/tendermint-rs/issues/741
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Duration(#[serde(with = "serializers::time_duration")] pub std::time::Duration);