+++ tendermint/src/public_key.rs
@@ -66,10 +66,17 @@ impl TryFrom<RawPublicKey> for PublicKey {
         let sum = &value
             .sum
             .ok_or_else(|| format_err!(error::Kind::InvalidKey, "empty sum"))?;
-        match sum {
-            Sum::Ed25519(b) => Self::from_raw_ed25519(b)
-                .ok_or_else(|| format_err!(error::Kind::InvalidKey, "malformed key").into()),
+        if let Sum::Ed25519(b) = sum {
+            return Self::from_raw_ed25519(b).ok_or_else(|| {
+                format_err!(error::Kind::InvalidKey, "malformed ed25519 key").into()
+            });
         }
+        #[cfg(feature = "secp256k1")]
+        if let Sum::Secp256k1(b) = sum {
+            return Self::from_raw_secp256k1(b)
+                .ok_or_else(|| format_err!(error::Kind::InvalidKey, "malformed key").into());
+        }
+        Err(format_err!(error::Kind::InvalidKey, "not an ed25519 key").into())
     }
 }
 
@@ -82,7 +89,11 @@ impl From<PublicKey> for RawPublicKey {
                 )),
             },
             #[cfg(feature = "secp256k1")]
-            PublicKey::Secp256k1(_) => panic!("secp256k1 PublicKey unimplemented"),
+            PublicKey::Secp256k1(ref pk) => RawPublicKey {
+                sum: Some(tendermint_proto::crypto::public_key::Sum::Secp256k1(
+                    pk.as_bytes().to_vec(),
+                )),
+            },
         }
     }
 }