+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   were Base64-encoded ([#717])
 - `[tendermint]` (Since v0.17.0-rc3) Bech32 encoding fix ([#690])
 - `[tendermint, light-client]` Specify the proposer in the validator set of fetched light blocks ([#705])
+- `[tendermint-proto]` (Since v0.17.0-rc3) Upgrade protobuf definitions to Tendermint Go v0.34.0 ([#737])
 
 [#425]: https://github.com/informalsystems/tendermint-rs/issues/425
 [#646]: https://github.com/informalsystems/tendermint-rs/pull/646
@@ -18,6 +19,7 @@
 [#701]: https://github.com/informalsystems/tendermint-rs/pull/701
 [#717]: https://github.com/informalsystems/tendermint-rs/issues/717
 [#705]: https://github.com/informalsystems/tendermint-rs/issues/705
+[#737]: https://github.com/informalsystems/tendermint-rs/pull/737
 
 ## v0.17.0-rc3